+++ mayan/apps/cabinets/models.py
@@ -1,7 +1,7 @@
 from __future__ import absolute_import, unicode_literals
 
 from django.core.exceptions import NON_FIELD_ERRORS, ValidationError
-from django.db import models, transaction
+from django.db import connection, models, transaction
 from django.urls import reverse
 from django.utils.encoding import python_2_unicode_compatible
 from django.utils.translation import ugettext_lazy as _
@@ -64,7 +64,12 @@ class Cabinet(MPTTModel):
         # https://code.djangoproject.com/ticket/1751
 
         with transaction.atomic():
-            if Cabinet.objects.select_for_update().filter(parent=self.parent, label=self.label).exists():
+            if connection.vendor == 'oracle':
+                queryset = Cabinet.objects.filter(parent=self.parent, label=self.label)
+            else:
+                queryset = Cabinet.objects.select_for_update().filter(parent=self.parent, label=self.label)
+
+            if queryset.exists():
                 params = {
                     'model_name': _('Cabinet'),
                     'field_labels': _('Parent and Label')